refactor: split shared modal controls
