+++ packages/ssrvpn_shared/lib/widgets/ssrvpn_node_selection_page.dart
@@ -11,6 +11,7 @@ import 'ssrvpn_app_surface.dart';
 part 'ssrvpn_node_selection_controls.dart';
 part 'ssrvpn_node_selection_keyboard.dart';
 part 'ssrvpn_node_selection_support_controls.dart';
+part 'ssrvpn_node_selection_subscription_filter.dart';
 part 'ssrvpn_node_selection_node_card.dart';
 
 typedef SsrvpnNodeAction = Future<void> Function(ProxyNode node);
